requirements.txt
certifi==2019.9.11
Django==2.2.7
djangorestframework==3.10.3
pytz==2019.3
sqlparse==0.3.0
